HWHoleSz must be u32...
Acked-by: Stefan Reinauer <stepan@coresystems.de>



git-svn-id: svn://svn.coreboot.org/coreboot/trunk@5430 2b7e53f0-3cfb-0310-b3e9-8179ed1497e1

+++ src/northbridge/amd/amdmct/mct/mctndi_d.c
@@ -33,7 +33,7 @@ void InterleaveNodes_D(struct MCTStatStruc *pMCTstat,
 	u8 DoIntlv;
 	u8 _NdIntCap;
 	u8 _SWHole;
-	u8 HWHoleSz;
+	u32 HWHoleSz;
 	u32 DramHoleAddrReg;
 	u32 HoleBase;
 	u32 dev0;
